:recycle: refactor: refatorando funções que podem ser úteis
