+++ src/components/viewer/viewer.tsx
@@ -8,7 +8,7 @@ import {
 	DialogTitle,
 	DialogTrigger
 } from "@/components/ui/dialog";
-import { DownloadFile } from "@/services/downloadFile";
+import { DownloadFile } from "@/services/utils/downloadFile";
 import { DownloadIcon } from "@heroicons/react/outline";
 import { useEffect, useState } from "react";
 import { DocumentViewer } from "react-documents";
