+++ src/services/myPatientService.tsx
@@ -12,20 +12,8 @@ export async function getData(id: string): Promise<Patient> {
 	return response.json();
 }
 
-export function formatPhone(phone: {
-	_ddi: string;
-	_ddd: string;
-	_number: string;
-}) {
-	return `${phone._ddi} (${phone._ddd}) ${phone._number}`;
-}
-
 export function formatMedicineSchedule(schedules: Date[]) {
 	return schedules.map((schedule, index) => {
 		return `${moment(schedule).format("HH:mm")} ${index !== schedules.length - 1 ? " / " : ""}`;
 	});
-}
-
-export function formatDateToddmmYYYY(date: Date) {
-	return date? moment(date).format("DD/MM/YYYY") : null;
 }
