+++ src/services/utils/formatDateToDDMMYYYY.ts
@@ -1,0 +1,5 @@
+import moment from "moment";
+
+export function formatDateToddmmYYYY(date: Date) {
+	return date? moment(date).format("DD/MM/YYYY") : null;
+}
